Add docket type badge to case details page
