+++ app/models/serializers/work_queue/appeal_serializer.rb
@@ -57,6 +57,10 @@ class WorkQueue::AppealSerializer < ActiveModel::Serializer
     object.advanced_on_docket
   end
 
+  attribute :docket_name do
+    object.docket_name
+  end
+
   attribute :docket_number do
     object.docket_number
   end
